Use MediaMetadata.Builder in MediaMetadataCompat.getMediaMetadata()
In MediaMetadataCompat.getMediaMetadata(), a parcel is used to parcel
the data from MediaMetadataCompat, and then the data is unparceled to
create a MediaMetadata object out of it.

This has an issue of creating intermediate ashmem bitmaps during the
parceling and unparceling depending on how many bitmaps are in the
metadata.  This can be avoided by passing the bitmap objects along
using MediaMedata.Builder without parceling/unparceling the bitmaps.
There are a few benefits of doing this:

  * it saves allocations of intermediate ashmem bitmaps
  * it avoids pixel copies between the bitmaps
  * it avoids creating an intermediate parcel

Bug: 400807118
Flag: EXEMPT refactor
Test: gradlew media:media:assemble
Test: gradlew media:media:test
Test: local test of youtube/youtube music with the change with no
intermediate ashmem bitmaps created

Change-Id: I41474740f137bc4bfe760efc6315753511acf6f2