Add a settings page
We have a button 'Disable overlay on background' that
appears on every demo page. In addition to taking up
space on every page, this also makes it hard to build
a focus demo with animated content that alternates
between a screen with focusable items and a screen
with no focusable items, because this settings button
is always visible and is focusable.

This CL adds a new settings page and moves the
'Disable overlay on background' setting to that page.

Bug: 413429406
Test: Demo only change, verified by running the demo.
Change-Id: I53588d2720cec1d569bb629fa98897ff9d5a03d4