Add credential transfer APIs
Introduce a new set of APIs for providers to enable transferring
credentials between the providers. The transfer requires the providers
to pre-register their export capabilities. When a user requests to
import credentials into a provider, a provider selector UI will be
launched with all the registered providers. After the user selects a
provider, the exporter's credential transfer activity will be launched
which will process the import request and return corresponding
credentials.

In order to avoid binder transaction limit, FileProvider is used as the
transfer medium. FileProvider provides a way for the providers transfer
credentials securely and reliably.

Relnote: "Added a new set of APIs for credential transfer. Credential
providers can now request credentials from another provider by
passing the ImportCredentialsRequest"

Test: unit tests
Change-Id: Idc79d77a585e759b5ac2f0a46cc56e22c053ef61