Update nested scroll interop connection.
In this CL I'm updating the nested scroll interop connection to better reflect the behavior in views. Previously, our connection would do the matching onPreFling -> onNestedPreFling and onPostFling -> onNestedFling but this is incorrect. In Views, onFling is called right before the fling is about to start on the child. Since we can't implement that direction in Scrollable we can use onPreFling for both onNestedPreFling and onNestedFling.

Relnote: Fixed a bug in NestedScrollInteropConnection where fling methods were being in the wrong order wrt to views.

Test: Added a new test.
Fixes: 416526536
Change-Id: I56ad4f04e5b29aceb7d7ceeb40976ad17aaf9ce0