Cache WindowAreaController when created
Removes the need to handle the scenario of
trying to perform an operation on a
WindowAreaInfo with a new WindowAreaController.

This will create a new WindowAreaController when
it is being created for the first time, thenreturn
the cached WindowAreaController anytime getOrCreate
is called after that.

This was originally merged and reverted in aosp/3246192,
added a test to verify the same object was being returned.

Bug: 390230656
Test: WindowAreaControllerTest
Test: WindowAreaControllerImplTest
Change-Id: I20588744a49246485675aca3b4814387b5f8e68d