Add UserInteractionTests class and a test to click on the remote content
The test uses the newly added hidden added fragment, to show remote
content in a scrollable container, and clicks on it.
It verifies the click by asserting that a Chrome activity is launched.
The test runs on the differet combination of:
- Mediation
- View/Compose
- Z-ordering

This change also moves the possible duplicate code of the integration
test classes to a newly added common parent class `AutomatedEndToEndTest`

Bug: 420602959
Bug: 402065627
Test: UserInteractionTests

Change-Id: Ice7e95bd4ede949186cdc68312abba1463f52c00