Refactor FocusOwner and FocusInvalidationManager
The FocusOwner and FocusInvalidationManager constructors
have a lot of lambda parameters, which result in unnecessary
Allocations. Over time we have added a lot of parameters and
the names of these parameters are confusing and the code is
hard to understand.

This CL introduces an internal interface that is passed as a
parameter to FocusOwner. AndroidComposeView implements
this interface that is passed to the FocusOwner constructor.

Similarly, all the lambdas needed by the FocusInvalidationManager
are replaced by functions added to FocusOwner.

Includes other minor code simplifications.

Bug: 413429406
Test: Ran all focus related tests
Change-Id: I4d022261a216e33b1f4fd6fe85892464d116e14c