Fix issue where value_type field of ranged value complication data is not copied over when creating WearSdkComplicationData.
This issue was happening when pushing a complication update from the provide app, which fitbit app was doing when initiating a sync from companion for example, among other use cases.

This was causing a rendering issue on 1p watch faces as the value type would change unexpectedly from 1 to 0 after pushing a new complication data value.

Bug: 417556239
Test: Manually verified that a recently built NightLight app that includes the fixed AndroidX library is working as intended, while the same build without this fix has this issue reproducible.
Change-Id: I759a9b58817ba06137512dbc53422d28196a25d7