Fix API lint errors in presentation layer
Hide APIs that aren't planned to be exposed publicly.
Hide APIs that are related to new, in-progress features.

Update existing @RestrictTo to apply to both the getter and the setter
for mutable properties.

Make nullability match platform methods for overrides, i.e. default to
nullable parameters where the Java code in the platform isn't
@NullMarked.

Avoid boxed Boolean in PdfView - use an enum to describe the "not set"
case for fastScrollVisibility.

Put non-default parameters first in PdfViewer.

Update Compose library to Kotlin-only. Compose itself can't be used in
Java code, and certain API lint checks are inapplicable to Kotlin-only
code.

Test: ./gradlew :pdf:pdf-viewer(test|cAT) and same for Compose
Change-Id: Id05d162961e0b36c1bef37b1314bf94412e63b52