Add all KMP targets supported by Annotations to required TestUtils
This change updates the TestUtils artifacts to include all KMP targets currently supported by the Annotations library. The targets added across the libraries include some or all of the following:

- JVM: Android, Desktop
- Native: Linux, iOS, watchOS, macOS, MinGW
- Web: JavaScript (JS), WasmJS

Note that targets will be added to the TestUtils library on an as-needed basis. I don't plan to update all TestUtils modules at once.

Test: N/A
Change-Id: Idf19e2560c09edd3b82e06768888d5f53c4a037a