Fix SpatialPopup's OutsideInputHandler
The InputCaptureView is capturing all touch/click events, even those that are technically in the target window. To work around this, I added a touch event listener to the target content inside of the OutsideInputHandler. This can then cancel the InputCaptureView's triggering of the onOutsideInput event.

Test: Ran Orbiter/Elevation test case
Bug: 417245722
Fixed: 417245722
Relnote: Fixed issue where SpatialPopup was being dismissed when clicked inside of the content.
Change-Id: If262c68747574f6cc986f23189ee2f67ad779af2