Make FillableData semantic properties and factories public
This changes makes relevant `fillableData` properties and factories public. It also makes the factories return nullable as the `actual` implementation of `FillableData` on Android requires API level 26 and above. Making the factories return nullable makes it safe to call on all API levels since it moves the version check inside the `actual` function body, returning a null object if autofill is not supported.

Test: Ensure existing tests pass
RelNote: Add semantic properties to set `FillableData` on components, and factories to create `FillableData` instances.
Change-Id: Icc5cf69cdec326bf9f12fdcb541d79e362793dfc