Add NavEntry.contentKey
NavEntry now requires a contentKey generator to generate a stable, unique saveable id for the entry's contentKey. This contentKey will sevrve as the primary id for the entry (an entry's content is to be identified by its contentKey).

As a result, NavKey.key will now be made private to enforce contentKey as the single source of truth for an entry's identity.

Test: n/a
Bug: 422001357
Bug: 420991203
Relnote: "NavEntry.key is now a private field. The NavEntry and its relevant states should be identified by the new contentKey parameter which is generated from the new contentKeyFactory lambda."
Change-Id: I81a6ced3a7bebf561c7206fcdf9402bc80c12752