Revert^2 "Handle SecurityException when opening cameras"
This is a reland of aosp/3650035 but also addresses an issue where
unnecessary parallel shutdown routines can lead to race condition and
deadlock, leading to CameraX shutdown hanging forever.

Here we remove the unnecessary path from `CameraInternal.release` to
close the individual camera. The path is only used for shutdown, and
`CameraPipe.shutdown` would shut down the cameras eventually.

Bug: 419602855
Bug: 422237649
Test: Affected integration tests
Change-Id: I8bf6f02df363611c568789387b1cdc6282d4a20d