Fix: Added sample for runComposeUiTest and modified documentation of SemanticsNodeInteraction.onChildren()
Previously, the documentation lacked specific information on how child nodes are captured from the tree for use in lazy lists. This ambiguity led to incorrect assumptions by developers, who expected a child to be present at a specific index simply because the data was passed to the list. This resulted in errors when accessing nodes that had not yet been created in the semantic tree due to its lazy-loading nature.

This commit updates the documentation to specify that the list will only fetch nodes that are currently present in the semantic tree. This provides developers with the necessary information to interact with the lazy list correctly and avoid errors.

Fixes: 225916436, 212279112
Test: Not Applicable
Change-Id: I1a7b60ca0136115a6e72f846435afdb21c4d7cc8