Change NavEntry.contentKey default value
Right now it defaults to NavEntry.key, which may not be saveable. However, we enforce contentKey to be saveable, so the default value we choose should be saveable. It now defaults to a hash on NavEntry.key instance.

Bug: 420991203
Bug: 422841812
Test: exisitng tests pass
Relnote: "NavEntry.contentKey now defaults to a saveable hash generated from NavEntry.key"
Change-Id: I2d7d4daba257e19bc9abe5ea64540126af18adf0