Release stale CameraInternal instances on re-initialization
During CameraRepository.init(), if an existing CameraInternal
for a given camera ID is being replaced by a new instance, the
old instance was not being released. This could lead to resource
leaks if init() is called multiple times.

This change checks the return value of Map.put(), which contains
the old value for the key. If this old value is non-null, we now
call release() on that stale instance to ensure its resources are
cleaned up before the new instance takes its place.

Bug: 423496824
Test: CameraXInitRetryTest
Change-Id: Id17b0ebf363e30e0820c46618258a6d21dc81ec7