Ignore material3 tests that are failing in postsubmit
As a result of moving to AGP-KMP, we are not able to set the targetSdk
for tests. This means that the targetSdk will be that of compileSdk,
which is 35 in material3's case.

Switching to 35 changes the default window insets:
https://developer.android.com/about/versions/15/behavior-changes-15#window-insets

We were accidentally not running these tests as they got dropped due to migration from legacy AGP plugin to the new AGP plugin in aosp/3642024

We now run these tests again since June 5, but they are failing and slowing down pre and post submit. Ignore for now

BUG: 422746273, 422735600, 422764314, 422746150, 422748931, 422764590, 422736884
Change-Id: Ic08d4c2ba0879a8a2301f21a9a2d9541b4b4528a