Migrate sql-bundled and uiautomator-shell to AGP KMP
This commit updates the `sqlite-bundled` and `uiautomator-shell` modules to use the new Kotlin Multiplatform Android library DSL (`androidLibrary { ... }`) instead of the older `androidTarget() { ... }` and separate `android { ... }` blocks.

This change also updates helper functions in `buildSrc` to work with the new `KotlinMultiplatformAndroidLibraryTarget` and `KotlinMultiplatformAndroidComponentsExtension` types.

BUG: 420602018
TESTED: Checked generated AAR and it contained the jni libs for the sqlite-bundled lib. Also ran ./development/validate_changes.sh and it passed
Change-Id: I568a10ab538d3afc1b9a1e840ec68b8f9fd6ca04