Fix OutputTransformation.addStyle crash in Material
Material TextFields deploy a trick to be able to run OutputTransformation independently to know its result. This is mainly used to understand whether the rendered text is going to be empty or not.

However, this trick obtains its TextFieldBuffer from a `TextFieldState.edit` call. Therefore, it is not initialized with an `OffsetMappingCalculator` that would come from an `OutputTransformation`. This triggers the deliberate crash we added to stop developers from calling `addStyle` from `TextFieldState.edit` or `InputTransformation.transformInput`.

This fix adds a smell detector to `TextFieldState.edit` to understand whether material TextField is calling it for the aforementioned purpose. In this case we let `OutputTransformation` call addStyle on the buffer.

Fix: b/424167352
Test: TextFieldTests in material and material3
Test: ./gradlew :compose:foundation:foundation:cAT
Change-Id: I6ee319022e87703205533ee4670947adb42997f4