Fix fullscreen dialog sizing issue on API 31 and below when decorFitsSystemWindows is true.
On API 31 and below, there is a bug in view framework (b/193978485)
where system bar insets were incorrectly considered to calculate
the max height of view resulting in view to be 1px or 2px
smaller than its parent.
To fix this issue we try to calculate the max height a dialog
can occupy after excluding system bar insets and set that
as target height.

Bug: b/211022812
Test: Existing tests should pass which were added earlier
Change-Id: I71b3b7ba631b0a4f71010bc72a62abf89d02f898