Actually add baseline-prof.txt to foundation AAR
When aosp/3653902 was reverted, it did not include reverting the location of baseline-prof.txt
KMP projects with AGP android plugin do not pacakge baseline-prof.txt if in a baselineProfiles directory. This is a bug (b/422388915).

I have a bug to create a check for this but it is not ready yet. Moving manually for now.

BUG: 22668628
Change-Id: Idd28e4281ac3160f1e1dcee8c9b0f65d06fc3cd4