Make BSTF follow the system setting
In the documentation of `TextObfuscationMode.RevealLastTyped`, we have claimed that this mode follows the system setting of showing password briefly but the implementation didn't. This change fixes this by employing a similar strategy as in `TextKeyListener` from platform.

This change also introduces a API-to-be `TextObfuscationMode.Default` that allows different platforms to choose different default obfuscation modes.

Finally, TextObfuscationMode documentation is updated to reduce vagueness.

Fix: 416410254
Fix: 416411014
Test: BasicSecureTextFieldTest
Change-Id: I41c0c91d93a8e4f5d3ae33fb6652997c50eb700c