Add API to disable metering repeating session
Originally if only an ImageCapture is bound, a meretring repeating will
be added internally to handle the focus metering controls.

ImageCapture.Builder.setMeteringRepeating allows users to disable the
metering repeating session explicitly, which can be useful if the extra
session is not supported on the device.

CameraControl.startFocusAndMetering will throw an excpetion when called
if only one ImageCapture is bound with metering repeating disabled.

Bug: b/423790236
Test: ./gradlew camera:camera-core:build,
./gradlew camera:camera-camera2:build

Change-Id: Ib06cf339f8376fdfd506982414afba4313890fd5