Clean up arguments for metalava compatibility baseline argument task
Inlines `getCommonBaselineUpdateArgs`, because it was only called from only place, and removes the `RestrictTo` arguments. These arguments are not needed because compatibility checks are done with signature files, not source files, and the signature files have already been filtered to the correct RestrictTo APIs (the `checkApiRelease` task does not use these arguments, and `ignoreApiChanges` should baseline the same errors raised by `checkApiRelease`).

Bug: 425907911
Test: Removed a `RestrictTo` method with `LIBRARY_GROUP_PREFIX` from core. Ran `./gradlew core:core:checkApi`, and verified that `./gradlew core:core:ignoreApiChanges` added the same error from `checkApi` to `core/core/api/restricted_current.ignore`
Change-Id: I67075f2cdae4cbe68fc802cae2972a945fa0cee9