Use AdHolder in PoolingContainer fragments
Instead of explicitly creating a SandboxedSdkView in the fragment and
corresponding layout, an AdHolder is used. It creates either a
SandboxedSdkView or a SharedUiContainer based on the chosen ad format,
and automatically fills it with content.

NATIVE_AD ad format is enabled for the CUJ as a part of the change.

The use of an AdHolder will help to add e2e PoolingContainer tests for
both SandboxedSdkView and SharedUiContainer at once (out of scope of
this change).

Test: build, manual
Bug: 422160411, Bug: 352500350
Change-Id: I7856318b9725bf88b347eecf2b613377de2fc095