Add CameraIdentifier and CameraSelector.of (Internal Use Only)
This change introduces CameraIdentifier and
its associated tests.

It's currently marked for library internal use only.
We plan to make it public once the broader external
camera modifications are complete.

Adds a new static factory method `CameraSelector.of(CameraIdentifier...)`
This is the new, recommended way to create a selector
for one or more specific cameras and it replaces the previous pattern.

Additionally, CameraInfo#getCameraIdentifier()
has been added, also marked for internal use for now.

Bug: 425891234
Test: CameraIdentifierTest
Change-Id: I732f0f52aceb45a93ae40b4bd0b6ef5e888603b2