Request both host and Projected device permissions when app requests camera or audio permission
For an app to access the Projected device camera or microphone, it needs both the host-scoped and Projected device-scoped permissions of camera or mic. Therefore, when an app requests a camera or mic permission using the Projected device permissions API, the request is split into two.

This change is part of a series introducing a complete permission request API surface.

Bug: b/418253909
Test: RequestProjectedPermissionsTest
Relnote: n/a
Change-Id: Ifc29e0c179c3b25541fdb01140569cd36f82b489