Introduce focusRect, controlled by FocusProperties
Custom focus rect is the adaptation of `View#getFocusedRect` in the compose world. Not many focusable layouts require a distinct internal area that defines where the focus should be. Text editors are one of the few exemptions. Focus should always follow the cursor or the selection rectangle in a TextField.

The new focusRect is also implemented via `FocusProperties` so it can be overridden by parent hierarchy. If someone wants to expand the area of the focus e.g. for KeepInView system, they can do it by adding new `focusProperties` modifier.

Fix: 430125252
Fix: 392901531
Bug: 368378073
Test: gradle :compose:ui:ui:cAT
Test: gradle :compose:foundation:foundation:cAT
Relnote: "Added customizable focus rect to `FocusProperties`. You can now define a custom focus area instead of defaulting the bounding box of the focusable. This information is used by the focus traversal system and the keep in view logic of scrollable containers."
Change-Id: Id6555453fe54b4a11db6a6ff90ebaa486b02f395