Hide @ExtensionFunctionType in java docs as well as kotlin
This annotation is no longer present if we switch dackka to k2, and the dokka team has said the issue is low priority since the annotation is internal and only used by the compiler. It won't really be useful to java developers anyway, so remove it from the docs now to minimize the diffs from k2.

Bug: 392589580
Test: ./gradlew docs
Change-Id: Ie329da3a3c0cd2b21535754deff1d04244130a82