Reset the SlotReader's slot array on close
There are cases where a SlotTable reference can retain a copy of
the slots array after it is closed that can cause it to hold onto
a old copy of the slot table (when the slot table grows) extending
the life of references longer than it should. The next composition
will reset this value but until the next composition the references
are retained.

Relnote: """Reduced the reference lifetime of values in the slot table
between compositions. There are cases where a reference in a slot table
could last a frame longer than necessary."""
Bug: 418516940
Test: Adhoc tests verify that the references are no longer retained.

Change-Id: I49e7488186782a95cb62358292b2950b5730089b