Fix ProtoLayoutDynamicDataPipelineTest
Due to the lack of built-in support of repeat delay in ValueAnimator, in our QuotaAwareAnimater we use animation.pause() inside the
onAnimationRepeat() and resume it later to achieve the repeat delay.

After the upgrade of robolectric from 4.14 to 4.15-beta, the end value
could not be reached before we paused the animation inside the
onAnimationRepeat().

To fix this broken test, we force the end value inside the
onAnimationRepeat before pause the animation. Also since the pause is
not intantanious stop, there is possible already calculated the next
frame's value and queued an onAnimationUpdate call for it, this value we
will ignore by checking animation.isPaused().

Bug: 424873228
Test: no more test failure
Change-Id: I7349eb830bb3407e44421434203236504bbe3759