Resolve UnsafeOptInUsageError in webkit
This CL adds the appropriate annotations to internal and external APIs
that were missing annotations.

The exception is WebViewFeatureInternal, which is expected to see a lot
of experimental API references, which is why the lint has been
suppressed for this file in particular.

Bug: 417458061
Test: ./gradlew webkit:webkit:lint
RelNote: """Adding experimental API annotations that had been previously
    omitted."""

Change-Id: Ic54eb7fc6176a032820f13cf46627d76197e2cc7