Add a FileDescriptor-based API for PdfLoader.
Not all PDFs come from content:// URIs. A FileDescriptor based API
unblocks apps that download PDFs from web services or load PDFs from
Files in their private storage from using the libary. Drive is such
an app.

This API still accepts a URI:
1) As a unique identifier. Bound services produce unique IBinders for
unique Intents. It's critical we have some way to provide a unique
Intent per document, and setting the data field is a good way to
accomplish this.
2) PdfDocument exposes Uri in its API. This is potentially worth
reconsidering, but it's not terrible to keep in place given #1.

Relnote: Adds FileDescriptor API to PdfLoader
Test: ./gradlew :pdf:pdf-document-service:cAT
Change-Id: I60b8dfa37b4c3800392f2d14992375b9d2581f05